+++ apps/api/src/main/java/br/org/apae/api/common/dto/paciente/dto/response/PatientResponseDTO.java
@@ -1,5 +1,6 @@
 package br.org.apae.api.common.dto.paciente.dto.response;
 
+import br.org.apae.api.common.dto.address.AddressResponseDTO;
 import br.org.apae.api.patient.domain.model.Patient;
 import java.time.LocalDate;
 import java.util.Collections;
